+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.3.0 (Jun 09, 2023)
+* Added Serverless VPC Access Connector and associated subnetwork.
+
+# 0.2.3 (May 10, 2023)
+* Enable Google Compute API.
+
 # 0.2.2 (Apr 20, 2023)
 * Increased default public and private subnets to larger ranges (`/16` instead of `/24`).